Great Leap Forward #1 continue

+++ euler/complete/prob_002.py
@@ -1,12 +1,11 @@
 # From: https://projecteuler.net/problem=2
-from mpmath import fibonacci
-from numpy import long
+from sympy import fibonacci
 
 if __name__ == "__main__":
-    total: long = 0
-    start_index: long = 3
+    total: int = 0
+    start_index: int = 3
 
-    while (current := long(fibonacci(start_index))) <= 4e6:
+    while (current := fibonacci(start_index)) <= 4e6:
         total += current
         start_index += 3
